Review fixes
- decommitUnusedPages: exclude every Structure subspace (BrandedStructure,
  WebAssemblyGCStructure too), skip empty blocks on the no-destructor path,
  assert the bit-vector lock discipline around isMarkingRetired, widen the
  page mask to 16 bits so 64 KB-block configurations (Linux arm64) work,
  include order.
- releaseUnusedSharedBaselineCode: no HeapIterationScope in the end phase
  (allocation is already stopped), reset the two-strike flag when an entry
  is dropped, and compile only under USE(BUN_JSC_ADDITIONS).
- The m_agedOut short-circuits and the unlinked-baseline-cache drop are
  USE(BUN_JSC_ADDITIONS)-only so a plain JSCOnly build keeps upstream's
  retention behaviour.

+++ Source/JavaScriptCore/heap/MarkedBlock.h
@@ -248,7 +248,7 @@ public:
             
         CellAttributes m_attributes;
         bool m_isFreeListed { false };
-        uint8_t m_decommittedPages { 0 }; // bit i set: OS page i of the block is decommitted
+        uint16_t m_decommittedPages { 0 }; // bit i set: OS page i of the block is decommitted
         unsigned m_index { std::numeric_limits<unsigned>::max() };
 
         AlignedMemoryAllocator* m_alignedMemoryAllocator { nullptr };
